### 2022-05-20 #4411 @editorialbot
[PRE REVIEW]: haldensify: Highly adaptive lasso conditional density estimation in R
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** https://github.com/nhejazi/haldensify
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.5
