+++ inst/REFERENCES.bib
@@ -5,7 +5,9 @@
     Longitudinal Studies},
   pages={167--180},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-319-65304-4_14},
+  url = {https://doi.org/10.1007/978-3-319-65304-4_14}
 }
 
 @article{diaz2012population,
@@ -17,7 +19,10 @@
   number={2},
   pages={541--549},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1111/j.1541-0420.2011.01685.x},
+  doi = {https://doi.org/10.1111/j.1541-0420.2011.01685.x}
+
 }
 
 @book{vdl2018targeted,
@@ -45,7 +50,9 @@
   number={1},
   pages={1--20},
   year={2011},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi = {10.2202/1557-4679.1356}
+  url = {https://doi.org/10.2202/1557-4679.1356}
 }
 
 @article{diaz2020causal,
@@ -69,7 +76,9 @@
     Longitudinal Studies},
   pages={77--94},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-319-65304-4_6},
+  url = {https://doi.org/10.1007/978-3-319-65304-4_6}
 }
 
 @article{vdl2017generally,
@@ -80,7 +89,9 @@
   volume={13},
   number={2},
   year={2017},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi = {10.1515/ijb-2015-0097},
+  url = {https://doi.org/10.1515/ijb-2015-0097}
 }
 
 @manual{coyle2022hal9001-rpkg,
@@ -111,7 +122,9 @@
   volume={2016},
   pages={689},
   year={2016},
-  organization={NIH Public Access}
+  organization={NIH Public Access},
+  doi = {10.1109/dsaa.2016.93},
+  url = {https://doi.org/10.1109/dsaa.2016.93}
 }
 
 @article{hirano2004propensity,
@@ -133,7 +146,9 @@
   number={1},
   pages={25--40},
   year={2015},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi = {10.1515/jci-2014-0022},
+  url = {https://doi.org/10.1515/jci-2014-0022}
 }
 
 @article{imai2004causal,
@@ -186,7 +201,9 @@
   Title = {Doubly robust estimation in missing data and causal inference
     models},
   Volume = {61},
-  Year = {2005}
+  Year = {2005},
+  doi = {10.1111/j.1541-0420.2005.00377.x},
+  url = {https://doi.org/10.1111/j.1541-0420.2005.00377.x}
 }
 
 @article{takeuchi2009nonparametric,
@@ -198,7 +215,9 @@
   number={2},
   pages={533--559},
   year={2009},
-  publisher={MIT Press}
+  publisher={MIT Press},
+  doi = {10.1162/neco.2008.10-07-628},
+  url = {https://doi.org/10.1162/neco.2008.10-07-628}
 }
 
 @article{bashtannyk2001bandwidth,
@@ -209,7 +228,9 @@
   number={3},
   pages={279--298},
   year={2001},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/s0167-9473(00)00046-3},
+  url = {https://doi.org/10.1016/s0167-9473(00)00046-3}
 }
 
 @inproceedings{neuneier1994estimation,
@@ -219,7 +240,9 @@
   booktitle={International Conference on Artificial Neural Networks},
   pages={689--692},
   year={1994},
-  organization={Springer}
+  organization={Springer},
[…]
