+++ inst/REFERENCES.bib
@@ -51,7 +51,7 @@
   pages={1--20},
   year={2011},
   publisher={De Gruyter},
-  doi = {10.2202/1557-4679.1356}
+  doi = {10.2202/1557-4679.1356},
   url = {https://doi.org/10.2202/1557-4679.1356}
 }
 
