+++ inst/REFERENCES.bib
@@ -312,7 +312,7 @@
   pages={131--154},
   year={2005},
   publisher={Elsevier},
-  doi = {10.1016/j.stamet.2005.02.003}
+  doi = {10.1016/j.stamet.2005.02.003},
   url = {https://doi.org/10.1016/j.stamet.2005.02.003}
 }
 
