### 2022-05-23 5c0907b
final bib tweaks for dois
