+++ inst/REFERENCES.bib
@@ -21,7 +21,7 @@
   year={2012},
   publisher={Wiley Online Library},
   doi = {10.1111/j.1541-0420.2011.01685.x},
-  doi = {https://doi.org/10.1111/j.1541-0420.2011.01685.x}
+  url = {https://doi.org/10.1111/j.1541-0420.2011.01685.x}
 
 }
 
