### 2022-05-23 #4411 @nhejazi
thanks @Kevin-Mattheus-Moerman, I've now added all of the missing DOIs and validated them against the checks from the editorial bot. no problem to wait until editors/reviewers are available.
