### 2022-06-24 #4411 @majensen
Hi @stefanocoretta - I appreciate your interest in reviewing for [JOSS](https://joss.theoj.org)! Are you available to review [this work](#4411)? Your expertise here would be very helpful; please have a look at the [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and if you are interested, please repond in this thread. Thanks for your consideration! --@majensen
