### 2022-06-24 #4411 @majensen
Hello  @turgeonmaxime - thanks for your interest in reviewing for [JOSS](https://joss.theoj.org). I believe your interests and expertise align with [this submitted work](#4411). Would you have time to review? Please have a look at [the review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and let us know by responding in this thread. If you have suggestions for other possible reviewers, that would be welcome as well. Many thanks for your consideration --@majensen
