### 2022-06-29 #4411 @stefanocoretta
> Hi @stefanocoretta - I appreciate your interest in reviewing for [JOSS](https://joss.theoj.org)! Are you available to review [this work](#4411)? Your expertise here would be very helpful; please have a look at the [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and if you are interested, please repond in this thread. Thanks for your consideration! --@majensen

Hello! Sorry for the late reply but I was away. I assume the review task has been taken over by someone already?
