### 2022-08-12 nhejazi/haldensify#37 @adibender
Example usage: 
Hi, this is regarding the software review here: https://github.com/openjournals/joss-reviews/issues/4522#4522

I like the vignette, but I think it could be improved a bit

- results plot should be described better. What do the colors indicate? What do we see? A legend that explains colors + dashed/solid lines etc. would probably be helpful
- would be nice to also see this applied to a real world example
