### 2022-08-12 nhejazi/haldensify#37 @adibender
Also, I ran the code and got the following output for the haldensify fit:

```
HAL Conditional Density Estimation
Number of bins over support of A: 30
CV-selected lambda: 4e-04
Summary of fitted HAL:
         coef                term
 1:  3.615390         (Intercept)
 2: -7.698355  [ I(W >= -0.102) ]
 3: -7.138447  [ I(W >= -0.402) ]
 4: -7.086183  [ I(W >= -2.771) ]
 5: -6.995051  [ I(W >= -3.578) ]
 6:  6.074327 [ I(bin_id >= 15) ]
 7: -6.008023  [ I(W >= -0.642) ]
 8: -5.901929  [ I(W >= -2.361) ]
 9:  5.856157  [ I(bin_id >= 9) ]
10:  5.763998   [ I(W >= 0.683) ]
Warning message:
In summary.hal9001(x$hal_fit) : Coefficients for many lambda exist --
Summarizing coefficients corresponding to minimum lambda.
```

What does the warning mean? Should I be worried? Readers/Users of the vignette will probably have similar questions. Maybe address it in the vignette.
