### 2022-08-12 #4522 @adibender
I've left an issue regarding the Example Usage vignette here: https://github.com/nhejazi/haldensify/issues/37
Otherwise good to go from my side
