### 2022-08-14 nhejazi/haldensify#37 @nhejazi
Thank you for your thorough review of this package @adibender! I've addressed the points above in https://github.com/nhejazi/haldensify/pull/38. The changes are summarized below

- I've added a paragraph below the plot in the vignette depicting the conditional density predictions against the theoretically correct densities for each of the groups. I found prose to be easier in getting the point across than a legend, the requisite concision of which could still omit important details.
- The major real-world use-case of this software is, as mentioned, for estimating the generalized propensity score in causal inference problems with continuous treatment variables. To that end, I've added a sentence saying that it is used in most applications of our 2020 _Biometrics_ paper. It is additionally used in the 2022 preprint described in the next section of the vignette.
- The warning is not a problem and, as indicated, is generated by `summary.hal9001`, a summary method called by `summary.haldensify`. I've suppressed this warning now in `summary.haldensify` as explaining this to a new user would be challenging, given that it requires a technical understanding of this algorithm and that implemented in the `hal9001` package.
