### 2022-08-15 #4522 @majensen
@adibender - can you indicate if @nhejazi 's addition at nhejazi/haldensify#38 meets your approval? If so, we can move towards acceptance. thanks!
