### 2022-08-23 #4522 @nhejazi
Hi @adibender — just another ping here about the outstanding issue for this open review. If you could comment on whether you’d like to see further edits or otherwise mark the issue as resolved, that would help us move forward.
