### 2022-08-24 nhejazi/haldensify#37 @adibender
Hi, thanks for the updates.
Some additional minor points.

- I've noticed you also added a section on IPW estimation. That's great, but makes the vignette somewhat lengthy. If it makes sense, consider splitting the vignette in two vignettes
- The references don't seem to render correctly in some cases? In some cases there's just a _____ in the beginning.
- Thanks for the references to the papers. In my experience for people interested in the software it's still helpful to see a concise walk through a real life example to get understanding of limitations, interpretation, etc. in absence of ground truth. But up to you.
