### 2022-08-24 nhejazi/haldensify#37 @nhejazi
thanks for the additional comments, a few responses and rationalizations below:

* I've noticed you also added a section on IPW estimation. That's great, but makes the vignette somewhat lengthy. If it makes sense, consider splitting the vignette in two vignettes
> The section on IPW estimation was actually already there, and, while I agree that the vignette is lengthier as a result, I'd rather have these two disjoint pieces presented together for a couple reasons. 1) A standalone vignette on IPW estimation would be quite short and would require the conditional density estimation material as prerequisite, so why not present it immediately after. 2) A user-side issue I've noticed with multiple vignettes is that users often only consult the first, so this would risk the IPW material being missed entirely. I'll keep this in mind and migrate the IPW material to a new section if I decide to expand this section significantly, which may yet happen since the statistical methodology paper on this is still only a preprint.

* The references don't seem to render correctly in some cases? In some cases there's just a _____ in the beginning.
> To the best of my knowledge, this is a stylistic choice made by the bibliographic management system that `pkgdown` uses under the hood (I'm not sure where the actual `.cls` file lives), but the underline/underscore is merely shorthand for indicating that the authorship team of the given article is the same as the preceding article. For example, you'll notice that in the rendered `README.md` (`index.html`), entries 5 and 6 in the bibliography display this, but this is because they are from the same set of authors as entry 4.

* Thanks for the references to the papers. In my experience for people interested in the software it's still helpful to see a concise walk through a real life example to get understanding of limitations, interpretation, etc. in absence of ground truth. But up to you.
> I agree with you here but feel that the vignette may be the wrong or a suboptimal place to present this information. My take is that it's much more useful to see real-world, possibly published examples of the methodology successfully applied, rather than including examples here for which context might be sparse. With coauthors, I've also been writing a to-be-published book that covers some of this material, with a section on conditional density estimation (https://tlverse.org/tlverse-handbook/sl3.html#pooled-hazard-regression), and I may add some more material there and link back here once we make further progress on that.

I appreciate your taking the time to review this thoroughly and will consider these finer points in improving the software package in the future.
