### 2022-08-24 #4522 @nhejazi
Update that the PR (https://github.com/nhejazi/haldensify/pull/38) has been merged after successfully passing CI checks.
