### 2022-09-03 #4522 @majensen
@nhejazi - I think the paper as such is very well written. What I am struggling with is, will a prospective user of the package be likely to read it? This may not be a big concern, given the popularity of the package. In the interest of making a better paper, though, I want to suggest making the internal paragraphs (i.e., not Statement of Need or Scope) 50-75% of their current length. This would force some choices about what previous work you want to review, and how much mathematical and statistical commentary you want to make.

Mathematical statistics is very far from my expertise. However, I can see that there is a flow in the presentation that starts with a problem to solve and leads logically to the solution you have created. Here is my precis of the paper (think of me as an NLP algorithm, not a statistician:)

> Conditional density estimation is a fundamental task across a number of areas of statistical analysis practice. It is frequently involved in the handling of nuisance parameters. Typically, parametric estimation of conditional densities for these parameters leads to poor quality model input, affecting quality of downstream estimation of the interesting parameters.
> Therefore, it is desirable to have a non-parametric method of conditional density estimation that can be rationally obtained using the data at hand. However, there are few implemented options for analysts to use.
>
> haldensify is an implementation of such an estimator, with desirable theoretical properties that make it highly suitable, e.g., for problems involving estimation of causal effect parameters in certain continuous treatment problems. In particular, effects of (time?) shifts in vaccine delivery on vaccine efficacy are amenable to estimation with our method.
>
> The non-parametric method of conditional density estimation of van der Laan and colleagues is the foundation of our implementation. Briefly, the method converts the density problem into a repeated-measures, hazard probability problem. The present work’s key advance is implementing this method using the highly adaptive lasso algorithm to perform the regression step on the transformed problem. The HAL algorithm has statistical properties that also enable us to rationally implement inverse probability weighted estimators based on the density estimation and (in our broad example class of problems) estimate causal effects.
>
> The haldensify package provides haldensify, the conditional density estimation algorithm. It also implements ipw_shift, which uses undersmoothing to select an appropriate generalized propensity score estimator (produced by haldensify) to obtain an inverse probability weighted estimator of causal effect with desirable properties. ipw_shift is another approach to the problem solved by txshift, which implements a so-called doubly robust estimator of the GPS.

This does a couple of things: it focuses on the implementation, leaves the details of the mathematical justifications and properties to references, and also puts your previous practical application at the beginning.

If this reduction is close to a correct interpretation, maybe it can help form a basis for a shorter paper. Let me know what you think. MAJ
