### 2022-09-04 #4522 @nhejazi
Thanks for weighing in on this and for the summary provided @majensen — I think what you’ve come up with is on point. I agree also that the readability — and brevity — of the paper is very important, much more so than the inclusion of possibly extraneous details. I’ll go through the paper and make cuts where I can later this week.
