### 2022-09-08 3db25b5
trim joss paper
