+++ paper/paper.md
@@ -3,10 +3,10 @@ title: "`haldensify`: Highly adaptive lasso conditional density estimation in `R
 tags:
   - machine learning
   - causal inference
-  - semiparametric estimation
   - conditional density estimation
   - generalized propensity score
   - inverse probability weighting
+  - semiparametric inference
   - R
 authors:
   - name: Nima S. Hejazi
@@ -27,7 +27,7 @@ affiliations:
     index: 3
   - name: Department of Biostatistics and Bioinformatics, Rollins School of Public Health, Emory University
     index: 4
-date: 24 August 2022
+date: 07 September 2022
 bibliography: ../inst/REFERENCES.bib
 ---
 
@@ -77,127 +77,81 @@ estimation of the generalized propensity score and useful for the construction
 of inverse probability weighted or doubly robust estimators of a class of causal
 effect parameters tailored to continuous treatments.
 
-# Flexible Estimation of Conditional Density Functions
+# Conditional Density Estimation and the Generalized Propensity Score
 
-Conditional density estimation is a challenging problem in statistical learning
-theory, significantly more complex than standard regression problems (i.e.,
-conditional mean estimation). Given the ubiquity of this functional parameter,
-it is unsurprising, then, that a range of techniques have been proposed,
-including kernel-based approaches [e.g., @bashtannyk2001bandwidth;
-@takeuchi2009nonparametric], conditional density ratio estimation [e.g.,
-@sugiyama2010conditional; @sugiyama2012density], and specialized neural network
-architectures [e.g., @neuneier1994estimation]. Yet, many of these approaches
-place restrictive assumptions on the density function (e.g., $k$-times
-differentiability), are computationally expensive or prohibitive (e.g., neural
-network models), avoid estimating the target quantity altogether (opting instead
-to estimate a density ratio), or fail to achieve convergence rates necessary for
-semiparametric estimation and inference.
+Conditional density estimation is a challenging and fundamental problem in
+statistical learning theory. Owing to the high frequency with which conditional
+density estimation arises in statistics and machine learning, a wide range of
+techniques have been proposed -- under a correspondingly wide range of
+assumptions. Some techniques are based in kernel smoothing [e.g.,
+@bashtannyk2001bandwidth; @takeuchi2009nonparametric], others in specialized
+neural network architectures [e.g., @neuneier1994estimation], and others still
+in the direct estimation of ratios of conditional densities [e.g.,
+@sugiyama2010conditional; @sugiyama2012density]. Most approaches make
+restrictive (parametric) assumptions about the form of the underlying density
+functional or fail to achieve convergence rates (of the estimator to the true,
+underlying conditional density) necessary for semiparametric inference. As such,
+analysts must often negotiate a difficult tradeoff between tractability, ease of
+implementation, and optimality properties of the chosen estimator. To partially
+resolve this open challenge, `haldensify` implements a nonparametric conditional
+density estimation procedure, making few assumptions regarding the underlying
+form of the density functional, with convergence-rate guarantees suitable for
+use in modern semiparametric inference and causal machine learning applications.
 
-Both @vdl2010targeted and @diaz2011super provided solutions to a subset of these
-issues by proposing a conditional density estimation algorithm that proceeds in
-three simple steps. First, the support of the variable of interest (e.g., the
-treatment) is partitioned into a user-specified number of bins, casting the
-dataset into a "long format" repeated measures structure, in which each unit is
-represented by a number of records (with the number of records matching the
-position of the bin into which the observed value of the variable of interest
-falls). Then, the hazard probability, conditional on any covariates, of the
[…]
