### 2022-09-08 5f4c4e3
trim joss paper more
