+++ paper/paper.md
@@ -56,77 +56,76 @@ estimators of causal effects [@diaz2012population; @haneuse2013estimation;
 
 Conditional density estimation is an important fundamental problem in the
 computational sciences and statistics, having garnered (independent) attention
-in machine learning [@takeuchi2009nonparametric; @sugiyama2010conditional;
-@sugiyama2012density], semiparametric estimation [@qin1998inferences;
-@cheng2004semiparametric], and causal inference [@hirano2004propensity;
-@vdl2010targeted; @diaz2011super; @zhu2015boosting]. Techniques for the
-nonparametric estimation of this quantity, complete with asymptotic optimality
-guarantees, have received comparatively limited attention. Similarly, despite
-the critical role of the generalized propensity score in the estimation of the
-causal effects of continuous treatments, this nuisance parameter is usually
-estimated with restrictive parametric modeling strategies, ultimately sharply
-limiting the quality of downstream point estimates and corresponding statistical
-inference (e.g., hypothesis tests, confidence intervals). Approaches for
-flexibly estimating the generalized propensity score have received limited
-attention [@diaz2011super; @zhu2015boosting], and software implementations of
-these techniques are, to the best of our knowledge, exceedingly rare, compared
-to, for example, regression algorithms for estimating conditional means.
-`haldensify` aims to partially fill this gap by implementing a flexible,
-nonparametric estimator of a conditional (or marginal) density, appropriate for
-estimation of the generalized propensity score and useful for the construction
-of inverse probability weighted or doubly robust estimators of a class of causal
-effect parameters tailored to continuous treatments.
+in machine learning [@takeuchi2009nonparametric; @sugiyama2012density],
+semiparametric estimation [@qin1998inferences; @cheng2004semiparametric], and
+causal inference [@hirano2004propensity; @diaz2011super; @zhu2015boosting].
+Techniques for the nonparametric estimation of this quantity, complete with
+asymptotic optimality guarantees, have received comparatively limited attention.
+Similarly, despite the critical role of the generalized propensity score in the
+estimation of the causal effects of continuous treatments, this nuisance
+parameter is usually estimated with restrictive parametric modeling strategies,
+ultimately sharply limiting the quality of downstream point estimates and
+corresponding statistical inference (e.g., hypothesis tests, confidence
+intervals). Approaches for flexibly estimating the generalized propensity score
+have received limited attention [@diaz2011super; @zhu2015boosting], and software
+implementations of these techniques are, to the best of our knowledge,
+exceedingly rare, compared to, for example, regression algorithms for estimating
+conditional means. `haldensify` aims to partially fill this gap by implementing
+a flexible, nonparametric estimator of a conditional (or marginal) density,
+appropriate for estimation of the generalized propensity score and useful for
+the construction of inverse probability weighted or doubly robust estimators of
+a class of causal effect parameters tailored to continuous treatments.
 
-# Conditional Density Estimation and the Generalized Propensity Score
+# Conditional Density Estimation and Modern Causal Inference
 
 Conditional density estimation is a challenging and fundamental problem in
 statistical learning theory. Owing to the high frequency with which conditional
 density estimation arises in statistics and machine learning, a wide range of
 techniques have been proposed -- under a correspondingly wide range of
 assumptions. Some techniques are based in kernel smoothing [e.g.,
-@bashtannyk2001bandwidth; @takeuchi2009nonparametric], others in specialized
-neural network architectures [e.g., @neuneier1994estimation], and others still
-in the direct estimation of ratios of conditional densities [e.g.,
-@sugiyama2010conditional; @sugiyama2012density]. Most approaches make
-restrictive (parametric) assumptions about the form of the underlying density
[…]
