### 2022-09-08 #4522 @nhejazi
Hi @majensen, an update to note that I've revised the paper draft to bring it more into line with your recommendations, which I agree with even more after having re-read the paper myself. My take is that the previous version of the paper included too much detail in the sections on methodology (on conditional density estimation and inverse probability weighted estimation); these same two sections have now been rewritten, removing much of the technical detail, and combined into a single section. Note that there now is some intentional overlap/redundancy that I hope reinforces concepts and the roles of particular user-facing functions within the `haldensify` package for the new reader. When you have a chance, please review and let me know what you think.
