### 2022-09-17 #4522 @majensen
@nhejazi I really appreciate your work on this. I think it reads very well and highlights the advance this algorithm makes within the context of the theory, whereas before I felt it was somewhat buried. It should now lead interested reader on to downloading and trying the software!

The final task is for you to create an open archive of the code repository, using Zenodo, FigShare or similar open service. A short tutorial on this is in this comment: https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729

Please respond back to this thread with the DOI of your archive. I will then add some metadata here and use the ancient spell summoning the Associate Editor in Chief!
