### 2022-09-17 #4522 @nhejazi
Thanks for confirming @majensen and glad to hear the updated paper reads better (I agree that the presentation of the key points is much improved in this draft).

Regarding the repository archive, the package/repository already had already been archived on Zenodo with a perpetual DOI (for all versions) of https://doi.org/10.5281/zenodo.3698329, so I've just created a new release of the package (https://github.com/nhejazi/haldensify/releases/tag/v0.2.5) and archived that updated version at https://doi.org/10.5281/zenodo.7089147. I think this should take care of the requirements for archiving the repository with a DOI.
