### 2022-09-22 #4522 @majensen
@nhejazi Can you review the above output from the paper build - I believe there are two references you can add ("Missing DOIs"), and it looks like there is a duplicate reference that could be deleted. Thanks!
