+++ inst/REFERENCES.bib
@@ -82,16 +82,16 @@
 }
 
 @article{vdl2017generally,
-  title={A Generally Efficient Targeted Minimum Loss Based Estimator based on
-    the Highly Adaptive Lasso},
+  title={A generally efficient targeted minimum loss-based estimator based on
+    the highly adaptive lasso},
   author={{van der Laan}, Mark J},
   journal={International Journal of Biostatistics},
   volume={13},
   number={2},
   year={2017},
-  publisher={De Gruyter},
   doi = {10.1515/ijb-2015-0097},
-  url = {https://doi.org/10.1515/ijb-2015-0097}
+  url = {https://doi.org/10.1515/ijb-2015-0097},
+  publisher = {De Gruyter}
 }
 
 @manual{coyle2022hal9001-rpkg,
@@ -164,8 +164,8 @@
 }
 
 @article{ertefaie2020nonparametric,
-  doi = {},
-  url = {https://arxiv.org/abs/2005.11303},
+  doi = {10.1111/biom.13719},
+  url = {https://doi.org/10.1111/biom.13719},
   year = {2022},
   publisher = {},
   journal = {Biometrics},
@@ -173,7 +173,7 @@
   number = {},
   pages = {},
   author = {Ertefaie, Ashkan and Hejazi, Nima S and {van der Laan}, Mark J},
-  title = {Nonparametric inverse probability weighted estimators based on the
+  title = {Nonparametric inverse-probability-weighted estimators based on the
     highly adaptive lasso}
 }
 
@@ -316,19 +316,6 @@
   url = {https://doi.org/10.1016/j.stamet.2005.02.003}
 }
 
-@article{vdl2017generally,
-  title={A generally efficient targeted minimum loss based estimator based on
-    the highly adaptive lasso},
-  author={{van der Laan}, Mark J},
-  journal={International Journal of Biostatistics},
-  volume={13},
-  number={2},
-  year={2017},
-  publisher={De Gruyter},
-  doi = {10.1515/ijb-2015-0097},
-  url = {https://doi.org/10.1515/ijb-2015-0097},
-}
-
 @article{bibaut2019fast,
   title={Fast rates for empirical risk minimization with cadlag losses with
     bounded sectional variation norm},
@@ -342,8 +329,14 @@
   title={Efficient estimation of pathwise differentiable target parameters with
     the undersmoothed highly adaptive lasso},
   author={{van der Laan}, Mark J and Benkeser, David and Cai, Weixin},
-  journal={arXiv preprint arXiv:1908.05607},
-  year={2019}
+  journal={International Journal of Biostatistics},,
+  doi = {10.1515/ijb-2019-0092},
+  url = {https://doi.org/10.1515/ijb-2019-0092},
+  year={2022},
+  volume = {(in press)},
+  number = {},
+  pages = {},
+  publisher = {De Gruyter}
 }
 
 @article{vdl2017uniform,
