### 2022-09-22 #4522 @nhejazi
@majensen --- thanks for catching these necessary edits. I've removed the duplicated bibliography entry and added the two missing DOIs (these manuscripts both completed their respective publication processes sometime after this software paper was submitted, so the DOIs were not available back then).
