### 2022-09-22 ed45134
catch bibtex typo
