+++ inst/REFERENCES.bib
@@ -329,7 +329,7 @@
   title={Efficient estimation of pathwise differentiable target parameters with
     the undersmoothed highly adaptive lasso},
   author={{van der Laan}, Mark J and Benkeser, David and Cai, Weixin},
-  journal={International Journal of Biostatistics},,
+  journal={International Journal of Biostatistics},
   doi = {10.1515/ijb-2019-0092},
   url = {https://doi.org/10.1515/ijb-2019-0092},
   year={2022},
